feat: ApplicationContext 구현
