+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -1,6 +1,7 @@
 package nextstep.mvc.tobe;
 
 import nextstep.db.DataBase;
+import nextstep.di.factory.ApplicationContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
@@ -15,7 +16,7 @@ public class AnnotationHandlerMappingTest {
 
     @BeforeEach
     public void setup() {
-        handlerMapping = new AnnotationHandlerMapping(TestConfiguration.class);
+        handlerMapping = new AnnotationHandlerMapping(new ApplicationContext(TestConfiguration.class));
         handlerMapping.initialize();
     }
